backend/user_authentication/serializers.py
from rest_framework import serializers
from .models import User
from django.contrib.auth import authenticate

class SignupSerializer(serializers.ModelSerializer):
    password = serializers.CharField(write_only=True, min_length=8)

    class Meta :
        model = User
        fields = ['email', 'name', 'password']

    def create(self, validated_data):
        email = validated_data['email']
        password = validated_data['password']
        name = validated_data['name']
        return User.objects.create_user(email=email, name=name, password=password)

class LoginSerializer(serializers.Serializer):
    email = serializers.EmailField()
    password = serializers.CharField(write_only=True, min_length=8)

    def validate(self, data):
        email = data['email']
        password = data['password']

        if email and password :
            user = authenticate(email=email, password=password)
            if not user:
                raise serializers.ValidationError("Invalid email or password")
            data["user"] = user
        else :
            raise serializers.ValidationError("Invalid email or password")

        return data
